+++ httpdlog/httpdlog-parser/src/test/java/nl/basjes/parse/httpdlog/NginxLogFormatTest.java
@@ -328,7 +328,7 @@ public class NginxLogFormatTest {
             .printAllPossibleValues();
     }
 
-    private class SingleFieldTestcase {
+    private static class SingleFieldTestcase {
         String logformat;
         String logline;
         String fieldName;